move `determine_next_face_on_edge` back to the `MeshWrapper`
